+++ MamaFit.BusinessObjects/Entity/MaternityDressDetail.cs
@@ -8,10 +8,10 @@ namespace MamaFit.BusinessObjects.Entity
         public string? Name { get; set; }
         public string? Description { get; set; }
         public string? Image { get; set; }
-        public decimal? Weight { get; set; }
+        public int? Weight { get; set; }
         public string? Color { get; set; }
         public string? Size { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
         
         // Navigation properties
